Refactor index view to remove unneeded methods to stores
